+++ gui_test/cypress/integration/AddLog/Happy Path.js
@@ -1,8 +1,6 @@
-//import {login} from './export/login.js';
+import {have_value, address} from '../../util/commands';
 
-const have_value = "have.value";
-
-describe("AddLogTask1",() => {
+describe("Happy Path",() => {
 
     const addLogBtn       = '//div[@id="root"]//li[1]//button[1]';
     const title           = '//input[@id="title"]';
@@ -35,7 +33,7 @@ describe("AddLogTask1",() => {
         const password        = '//input[@id="password"]';
         const loginBtn        = '//input[@id="kc-login"]';
 
-        cy.visit("https://keycloak-beta.hsiang.me/auth/realms/OIS/protocol/openid-connect/auth?client_id=timelog&redirect_uri=http%3A%2F%2Flocalhost%3A3000%2F&state=767eb54d-bccd-4fa2-8cca-2379befcb658&response_mode=fragment&response_type=code&scope=openid&nonce=d01888f5-cf4f-4dd5-8d24-d095b6d254e0&code_challenge=tMRafhLcHYLVT8oep70S8eHvp9-eB548bB4Cl9OeZvg&code_challenge_method=S256");
+        cy.visit(address);
         //-------------login---------------
         //username
         cy.xpath(userName)
