+++ gui_test/cypress/integration/AddLog/Reversed date.js
@@ -1,7 +1,7 @@
-const have_value = "have.value";
+import {have_value, address} from '../../util/commands';
 
-//Empty Activity
-describe("AddLogTask3",() => {
+//Reversed date
+describe("Reversed date",() => {
 
     const hourAM          = '//span[@class="MuiButton-label"]/h6[text()="AM"]';
     const hourPM          = '//span[@class="MuiButton-label"]/h6[text()="PM"]';
@@ -19,7 +19,7 @@ describe("AddLogTask3",() => {
         const userID          = "ssl1321ois";
         const userPassword    = "lab1321bal";
 
-        cy.visit("https://keycloak-beta.hsiang.me/auth/realms/OIS/protocol/openid-connect/auth?client_id=timelog&redirect_uri=http%3A%2F%2Flocalhost%3A3000%2F&state=767eb54d-bccd-4fa2-8cca-2379befcb658&response_mode=fragment&response_type=code&scope=openid&nonce=d01888f5-cf4f-4dd5-8d24-d095b6d254e0&code_challenge=tMRafhLcHYLVT8oep70S8eHvp9-eB548bB4Cl9OeZvg&code_challenge_method=S256");
+        cy.visit(address);
         
         cy.login(userID,userPassword);
     });
@@ -37,7 +37,7 @@ describe("AddLogTask3",() => {
           .type("Self Reading")
           .should(have_value,"Self Reading");
     });
-    /*
+
     it("Add activity type",()=>{
         const activityType    = '//div[@id="activity-type-select"]';
         const labProject      = '//li[contains(text(),"LabProject")]';
@@ -49,13 +49,12 @@ describe("AddLogTask3",() => {
             expect($div).to.have.text('LabProject')
         });
     });
-    */
 
-   it("Select start date",()=>{
+    it("Select start date",()=>{
         const startDateBox    = '//div[@class="MuiPaper-root MuiDialog-paper MuiDialog-paperScrollPaper MuiDialog-paperWidthSm MuiPaper-elevation24 MuiPaper-rounded"]//div[1]//div[1]//div[1]//div[1]//input[1]';
         const startYear       = '//div[contains(text(),"' + yyyy + '")]';
         const startMonth      = '//p[@class="MuiTypography-root MuiTypography-body1 MuiTypography-alignCenter"]'
-        const okBtn           = '//span[contains(text(),"OK")]';
+        const okBtn = '//span[contains(text(),"OK")]';
         
         cy.ClickTo(startDateBox);
         
@@ -85,7 +84,7 @@ describe("AddLogTask3",() => {
         //cy.log(yyyy + "/" + mm + "/" + dd);
         //cy.ClickTo(date_1);
         cy.xpath(startDateBox)
-        .should(have_value,yyyy + "/" + mm + "/" + dd);
+          .should(have_value,yyyy + "/" + mm + "/" + dd);
     });
 
     it("Select start time",()=>{
@@ -95,19 +94,19 @@ describe("AddLogTask3",() => {
         cy.ClickTo(startTimeBox);
         //AM
         cy.ClickTo(hourAM);
-        //9
-        cy.xpath(clock).click(10,130);
+        //7
+        cy.xpath(clock).click(74,230);
         //00
         cy.xpath(clock).click(130,10);
         cy.xpath(startTimeBox)
-          .should(have_value,"09:00 AM");
+          .should(have_value,"07:00 AM");
     });
     
     it("Select end date",()=>{
         const endDateBox      = '//div[@class="MuiGrid-root MuiGrid-container MuiGrid-spacing-xs-3"]//div[3]//div[1]//div[1]//div[1]//input[1]';
-        const endYear         = '//div[contains(text(),"' + yyyy +'")]';
-        const endMonth        = '//p[@class="MuiTypography-root MuiTypography-body1 MuiTypography-alignCenter"]';
-        const okBtn           = '//span[contains(text(),"OK")]';
+        const endYear       = '//div[contains(text(),"2019")]';
+        const endMonth      = '//p[@class="MuiTypography-root MuiTypography-body1 MuiTypography-alignCenter"]';
+        const okBtn = '//span[contains(text(),"OK")]';
         
         //end date
         cy.ClickTo(endDateBox);
@@ -131,7 +130,7 @@ describe("AddLogTask3",() => {
         //     while(j < 12);  
         // })
         cy.xpath(endDateBox)
-          .should(have_value, yyyy + "/" + mm + "/" + dd);
+          .should(have_value,"2019/" + mm + "/" + dd);
     });
     
     it("Select end time",()=>{
@@ -166,7 +165,7 @@ describe("AddLogTask3",() => {
 
         //alert box
         cy.on('window:alert',(str)=>{
-            expect(str).to.equal('Activity Type is not selected.')
+            expect(str).to.equal('Start Time should be earlier than End Time.')
         });
     });
 });
