+++ gui_test/cypress/integration/Login/Empty Password.js
@@ -1,0 +1,29 @@
+import {have_value, address} from '../../util/commands.js'
+
+describe("Empty Password", () =>{
+    
+    it("Login to the website", ()=>{
+        const userID          = "ssl1321ois";
+        const userName        = '//input[@id="username"]';
+        const loginBtn        = '//input[@id="kc-login"]';
+
+        cy.visit(address);
+        
+        //username
+        cy.xpath(userName)
+         .type(userID)
+         .should(have_value, userID); 
+        
+         //click login
+         cy.ClickTo(loginBtn);
+    });
+
+    it("Assert Feedback", () =>{
+        const feedbackPath    = '//span[@class="kc-feedback-text"]';
+
+        cy.xpath(feedbackPath).should(($el) =>{
+            expect($el).to.have.text("Invalid username or password.");
+        });
+    });
+    
+})
