+++ gui_test/cypress/util/commands.js
@@ -1,0 +1,8 @@
+const have_value = "have.value";
+const have_attr  = "have.attr";
+const have_text  = "have.text";
+
+const address    = "https://keycloak-beta.hsiang.me/auth/realms/OIS/protocol/openid-connect/auth?client_id=timelog&redirect_uri=http%3A%2F%2Flocalhost%3A3000%2F&state=767eb54d-bccd-4fa2-8cca-2379befcb658&response_mode=fragment&response_type=code&scope=openid&nonce=d01888f5-cf4f-4dd5-8d24-d095b6d254e0&code_challenge=tMRafhLcHYLVT8oep70S8eHvp9-eB548bB4Cl9OeZvg&code_challenge_method=S256";
+
+
+export {have_value, have_attr, have_text, address};
